Change the control flow logic for import requests
