+++ iamctl/pkg/userStores/import.go
@@ -19,22 +19,14 @@
 package userstores
 
 import (
-	"bytes"
-	"crypto/tls"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
-	"mime"
-	"mime/multipart"
-	"net/http"
-	"net/textproto"
 	"os"
 	"path/filepath"
 	"strings"
 
 	"github.com/wso2-extensions/identity-tools-cli/iamctl/pkg/utils"
-	"gopkg.in/yaml.v2"
 )
 
 func ImportAll(inputDirPath string) {
@@ -72,6 +64,7 @@ func ImportAll(inputDirPath string) {
 
 func importUserStore(userStoreId string, importFilePath string) error {
 
+	fmt.Println(userStoreId)
 	fileBytes, err := ioutil.ReadFile(importFilePath)
 	if err != nil {
 		return fmt.Errorf("error when reading the file for user store: %s", err)
@@ -82,114 +75,16 @@ func importUserStore(userStoreId string, importFilePath string) error {
 	userStoreKeywordMapping := getUserStoreKeywordMapping(fileInfo.ResourceName)
 	modifiedFileData := utils.ReplaceKeywords(string(fileBytes), userStoreKeywordMapping)
 
-	err = sendImportRequest(userStoreId, importFilePath, modifiedFileData)
+	if userStoreId == "" {
+		log.Println("Creating new user store: " + fileInfo.ResourceName)
+		err = utils.SendImportRequest(importFilePath, modifiedFileData, utils.USERSTORES)
+	} else {
+		log.Println("Updating user store: " + fileInfo.ResourceName)
+		err = utils.SendUpdateRequest(userStoreId, importFilePath, modifiedFileData, utils.USERSTORES)
+	}
 	if err != nil {
 		return fmt.Errorf("error when importing user store: %s", err)
 	}
+	log.Println("User store imported successfully.")
 	return nil
-}
-
-func sendImportRequest(userStoreId string, importFilePath string, fileData string) error {
-
-	fileInfo := utils.GetFileInfo(importFilePath)
-
-	var requestMethod, reqUrl string
-	if userStoreId != "" {
-		log.Println("Updating user store: " + fileInfo.ResourceName)
-		reqUrl = utils.SERVER_CONFIGS.ServerUrl + "/t/" + utils.SERVER_CONFIGS.TenantDomain + "/api/server/v1/userstores/" + userStoreId + "/import"
-		requestMethod = "PUT"
-	} else {
-		log.Println("Creating new user store: " + fileInfo.ResourceName)
-		reqUrl = utils.SERVER_CONFIGS.ServerUrl + "/t/" + utils.SERVER_CONFIGS.TenantDomain + "/api/server/v1/userstores/import"
-		requestMethod = "POST"
-	}
-
-	var buf bytes.Buffer
-	var err error
-	_, err = io.WriteString(&buf, fileData)
-	if err != nil {
-		return fmt.Errorf("error when creating the import request: %s", err)
-	}
-
-	mime.AddExtensionType(".yml", "application/yaml")
-	mime.AddExtensionType(".xml", "application/xml")
-	mime.AddExtensionType(".json", "application/json")
-
-	mimeType := mime.TypeByExtension(fileInfo.FileExtension)
-
-	body := new(bytes.Buffer)
-	writer := multipart.NewWriter(body)
-	defer writer.Close()
-
-	part, err := writer.CreatePart(textproto.MIMEHeader{
-		"Content-Disposition": []string{fmt.Sprintf(`form-data; name="%s"; filename="%s"`, "file", fileInfo.FileName)},
-		"Content-Type":        []string{mimeType},
-	})
-	if err != nil {
-		return fmt.Errorf("error when creating the import request: %s", err)
-	}
-
-	_, err = io.Copy(part, &buf)
-	if err != nil {
-		return fmt.Errorf("error when creating the import request: %s", err)
-	}
-
-	request, err := http.NewRequest(requestMethod, reqUrl, body)
-	request.Header.Add("Content-Type", writer.FormDataContentType())
-	request.Header.Set("Authorization", "Bearer "+utils.SERVER_CONFIGS.Token)
-	defer request.Body.Close()
-
-	if err != nil {
-		return fmt.Errorf("error when creating the import request: %s", err)
-	}
-	client := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				InsecureSkipVerify: true,
-			},
-		},
-	}
-	resp, err := client.Do(request)
-	if err != nil {
-		return fmt.Errorf("error when sending the import request: %s", err)
-	}
-
-	statusCode := resp.StatusCode
-	if statusCode == 201 {
-		log.Println("User store created successfully.")
-		return nil
-	} else if statusCode == 200 {
-		log.Println("User store updated successfully.")
-		return nil
-	} else if statusCode == 409 {
-		log.Println("A user store with the same name already exists.")
-	} else if error, ok := utils.ErrorCodes[statusCode]; ok {
-		return fmt.Errorf("error response for the import request: %s", error)
-	}
-	return fmt.Errorf("unexpected error when importing user store: %s", resp.Status)
-}
-
-func getUserStoreId(userStoreFilePath string) (string, error) {
-
-	fileContent, err := ioutil.ReadFile(userStoreFilePath)
-	if err != nil {
-		return "", fmt.Errorf("error when reading the file: %s. %s", userStoreFilePath, err)
-	}
-	var userStoreConfig UserStoreConfigurations
-	err = yaml.Unmarshal(fileContent, &userStoreConfig)
-	if err != nil {
-		return "", fmt.Errorf("invalid file content at: %s. %s", userStoreFilePath, err)
-	}
-
-	existingUserStoreList, err := getUserStoreList()
-	if err != nil {
-		return "", fmt.Errorf("error when retrieving the deployed userstore list: %s", err)
-	}
-
-	for _, userstore := range existingUserStoreList {
-		if userstore.Id == userStoreConfig.ID {
-			return userstore.Id, nil
-		}
-	}
-	return "", nil
 }
